usher/requirements.txt
httpx>=0.25.0
pyyaml>=6.0
click>=8.1.0
# Linux/Raspberry Pi only (IR input):
# evdev>=1.6.0
